imagebitmap-renderingcontext/toBlob-origin-clean-offscreen.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>createImageBitmap: origin-clean flag</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<script src="/common/namespaces.js"></script>
<div id=log></div>
<body>
<script>

async_test(function(t) {
    var image = document.createElement('img');
    image.src = "http://{{domains[www1]}}:{{ports[http][0]}}/images/red.png";
    image.addEventListener('load', function() {
        var offscreenCanvas = new OffscreenCanvas(10, 10);
        var contextBitmap = offscreenCanvas.getContext('bitmaprenderer');
        createImageBitmap(img, 0, 0, img.width, img.height).then(function(bitmap) {
          contextBitmap.transferFromImageBitmap(bitmap);
          offscreenCanvas.convertToBlob().then(t.step_func_done(function() {
              assert_false("convertToBlob on a tainted OffscreenCanvas didn't throw, but should be");
          }), t.step_func_done(function(e) {
              assert_true(e instanceof DOMException);
              assert_equals(e.name, "SecurityError");
          }));
        });
    });
}, "Test that call convertToBlob on a tainted OffscreenCanvas throws exception");


</script>
</body>
